New test cases
I’ve done 8 test. Divided in 3 big areas: bitrate, type and list.

In bitrate I separate between max Bitrate in “sameBitrate_done”  is checking a test where 2 of the 4rd videos are with the same Bitrate. In “plus_1_bitrate_done”  is checking the that one of the video has just 1 bitrate more than the other. “less_1_bitrate_done” is checking the opposite case of the  “plus_1_bitrate_done”. “bitrate_negative_done” is checking where one of those videos is negative. Last one in this area is “NaN_mp4” is checking where one of the videos don’t have bitrate instead of one that have it.

The next area is type where the first one is “not_MP4_done” were either of the videos are MP4. The “max_Bitrate_notMP4_done”  when one has the biggest bitrate but is not an MP4 video. 

Last area is list where there is no video list.